+++ gitops/git/bitbucket/BUILD
@@ -15,7 +15,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["bitbucket.go"],
-    importpath = "github.com/adobe/rules_gitops/gitops/vcs/bitbucket",
+    importpath = "github.com/adobe/rules_gitops/gitops/git/bitbucket",
     visibility = ["//visibility:public"],
 )
