+++ gitops/git/github/BUILD
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["github.go"],
-    importpath = "github.com/adobe/rules_gitops/gitops/vcs/github",
+    importpath = "github.com/adobe/rules_gitops/gitops/git/github",
     visibility = ["//visibility:public"],
     deps = [
         "//vendor/github.com/google/go-github/v32/github:go_default_library",
