+++ gitops/prer/BUILD
@@ -23,9 +23,7 @@ go_library(
         "//gitops/commitmsg:go_default_library",
         "//gitops/exec:go_default_library",
         "//gitops/git:go_default_library",
-        "//gitops/vcs:go_default_library",
-        "//gitops/vcs/bitbucket:go_default_library",
-        "//gitops/vcs/github:go_default_library",
+        "//gitops/git/bitbucket:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
     ],
 )
