--- gitops/vcs/BUILD
@@ -1,8 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["vcs.go"],
-    importpath = "github.com/adobe/rules_gitops/gitops/vcs",
-    visibility = ["//visibility:public"],
-)
